+++ tests/cli.rs
@@ -21,7 +21,7 @@ struct Smoke {
     report_name: String,
 }
 
-#[derive(Clone, Copy, cucumber::World, Debug, Default)]
+#[derive(Clone, Copy, Debug, Default, cucumber::World)]
 struct World;
 
 #[given("an invalid step")]