+++ tests/filter.rs
@@ -40,7 +40,7 @@ async fn by_examples() {
     }
 }
 
-#[derive(Clone, Copy, cucumber::World, Debug, Default)]
+#[derive(Clone, Copy, Debug, Default, cucumber::World)]
 struct World;
 
 /// Deterministic output of [`writer::Basic`].