+++ src/writer/summarize.rs
@@ -539,9 +539,7 @@ pub trait Summarizable {}
 
 impl<T: writer::NonTransforming> Summarizable for T {}
 
-// TODO: Try remove on next Rust version update.
-#[expect(clippy::allow_attributes, reason = "`#[expect]` doesn't work here")]
-#[allow( // intentional
+#[expect( // related to summarization only
     clippy::multiple_inherent_impl,
     reason = "related to summarization only"
 )]