+++ codegen/CHANGELOG.md
@@ -13,10 +13,10 @@ All user visible changes to `cucumber-codegen` crate will be documented in this 
 
 ### BC Breaks
 
-- Bumped up [MSRV] to 1.81 to get rid of `once_cell` crate and for `#[expect]` attribute usage. ([4010c1ad], [todo])
+- Bumped up [MSRV] to 1.81 to get rid of `once_cell` crate and for `#[expect]` attribute usage. ([4010c1ad], [f1307038])
 
 [4010c1ad]: /../../commit/4010c1ad6a53d6b7f0b28cefea73c8c13e880e9f
-[todo]: /../../commit/todo-full
+[f1307038]: /../../commit/f1307038cb6b1e38c1cc259a0e09fb583033d0cf